rename enum-service
